Rename utils.js to package-rollup.js
